Final Version

+++ code/CourseManagementSystem/src/view/commands/GradeStudentCommand.java
@@ -3,8 +3,8 @@ package view.commands;
 public class GradeStudentCommand implements Command {
 	private String studentName;
 	private String courseName;
-	private int grade;
-	public GradeStudentCommand(String studentName, String courseName, int grade) {
+	private double grade;
+	public GradeStudentCommand(String studentName, String courseName, double grade) {
 		this.studentName = studentName;
 		this.courseName = courseName;
 		this.grade = grade;
@@ -15,7 +15,7 @@ public class GradeStudentCommand implements Command {
 	public String getCourseName() {
 		return courseName;
 	}
-	public int getGrade() {
+	public double getGrade() {
 		return grade;
 	}
 }
